I added password protection to the server, though right now the password is not configurable and it's "passw0rd".
Also, I used the data value of a monster spawner to record the I.D. of the monster it spawns instead of simply relying on the name of the spawner item to determine what it spawns; therefore, people will not longer be able to use anvils to change spawners to different spawners. This new system has not yet been tested, however.